# i-guest-house-ui/src/open-api/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api/v1",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/rooms/{id}": {
            "put": {
                "tags": [
                    "room-controller"
                ],
                "operationId": "updateRoom",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Room"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "room-controller"
                ],
                "operationId": "deleteRoom",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/booking/update-room-number": {
            "put": {
                "tags": [
                    "booking-controller"
                ],
                "operationId": "updateRoomNumber",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/RoomUpdateRequest"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rooms": {
            "get": {
                "tags": [
                    "room-controller"
                ],
                "operationId": "getAllRooms",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Room"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "room-controller"
                ],
                "operationId": "createRoom",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Room"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/food/preference": {
            "get": {
                "tags": [
                    "food-preference-controller"
                ],
                "operationId": "getAllFoodPreference",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FoodPreferenceDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "food-preference-controller"
                ],
                "operationId": "createFoodPreference",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FoodPreferenceDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dailyMeal": {
            "get": {
                "tags": [
                    "daily-meal-controller"
                ],
                "operationId": "getMeal",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DailyMeal"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "daily-meal-controller"
                ],
                "operationId": "insertMeal",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DailyMeal"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/booking": {
            "get": {
                "tags": [
                    "booking-controller"
                ],
                "operationId": "getAllBookings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingRequest"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "booking-controller"
                ],
                "operationId": "createBookingRequest",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/BookingRequest"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingRequest"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "authentication-controller"
                ],
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "authentication-controller"
                ],
                "operationId": "authenticate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthenticationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/booking/user": {
            "get": {
                "tags": [
                    "booking-controller"
                ],
                "operationId": "getBookingHistory",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingHistory"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/booking/roomType/{roomType}": {
            "get": {
                "tags": [
                    "booking-controller"
                ],
                "operationId": "getBookingRequestsByRoomType",
                "parameters": [
                    {
                        "name": "roomType",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingHistory"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/booking/notAlloted": {
            "get": {
                "tags": [
                    "booking-controller"
                ],
                "operationId": "getAllNotAllotedBookings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingRequest"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/booking/HR": {
            "get": {
                "tags": [
                    "booking-controller"
                ],
                "operationId": "getBookingsOfHR",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingRequest"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Room": {
                "type": "object",
                "properties": {
                    "roomId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "roomNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "roomType": {
                        "type": "string"
                    },
                    "roomCapacity": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "roomCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isAvailable": {
                        "type": "boolean"
                    },
                    "floorNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "towerNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "furnitures": {
                        "type": "string"
                    }
                }
            },
            "RoomUpdateRequest": {
                "type": "object",
                "properties": {
                    "bookingId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "roomId": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "FoodPreferenceDto": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "string"
                    },
                    "userName": {
                        "type": "string"
                    },
                    "breakfast": {
                        "type": "boolean"
                    },
                    "lunch": {
                        "type": "boolean"
                    },
                    "dinner": {
                        "type": "boolean"
                    },
                    "preferences": {
                        "type": "string"
                    },
                    "mealDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "DailyMeal": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "breakfast": {
                        "type": "string"
                    },
                    "lunch": {
                        "type": "string"
                    },
                    "dinner": {
                        "type": "string"
                    },
                    "mealDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "BookingRequest": {
                "required": [
                    "bookedToId",
                    "checkInDate",
                    "checkOutDate",
                    "guestName",
                    "roomType"
                ],
                "type": "object",
                "properties": {
                    "bookingId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "roomNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "guestName": {
                        "type": "string"
                    },
                    "bookedToId": {
                        "type": "string"
                    },
                    "roomType": {
                        "type": "string"
                    },
                    "checkInDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "checkOutDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "bookedDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "UserRequest": {
                "required": [
                    "assoId",
                    "assoLevel",
                    "email",
                    "firstName",
                    "lastName",
                    "mobileNumber",
                    "password",
                    "roleId"
                ],
                "type": "object",
                "properties": {
                    "assoId": {
                        "type": "string"
                    },
                    "assoLevel": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "mobileNumber": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "roleId": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "AuthenticationResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    }
                }
            },
            "AuthenticationRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "maxLength": 2147483647,
                        "minLength": 8,
                        "type": "string"
                    }
                }
            },
            "BookingHistory": {
                "type": "object",
                "properties": {
                    "bookingId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "checkInDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "checkOutDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "roomNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "floorNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "towerNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "bookedDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "userName": {
                        "type": "string"
                    }
                }
            }
        }
    }
}